refactor: Rename `lib/wrapClean.js`
