+++ lib/plugin.js
@@ -10,7 +10,7 @@ const logout = require('./logout');
 const wrap = require('./wrap');
 const injectLogsIamRole = require('./inject-logs-iam-role');
 const injectOutputOutputs = require('./inject-output-outputs');
-const wrapClean = require('./wrapClean');
+const wrapClean = require('./wrap-clean');
 const getAppUids = require('./app-uids');
 const removeDestination = require('./remove-destination');
 const { saveDeployment, createAndSetDeploymentUid } = require('./deployment');
