+++ lib/plugin.test.js
@@ -45,7 +45,7 @@ describe('plugin', () => {
     ServerlessEnterprisePlugin = proxyquire('./plugin', {
       './logs-collection': logsCollection,
       './wrap': wrap,
-      './wrapClean': wrapClean,
+      './wrap-clean': wrapClean,
       './remove-destination': removeDestination,
       './deployment': { saveDeployment, createAndSetDeploymentUid },
       './generate-event': { eventDict: {}, generate },
