+++ lib/wrap-clean.js
