+++ lib/wrap-clean.test.js
@@ -13,7 +13,7 @@ describe('wrapClean', () => {
       pathExistsSync: sinon.stub().returns(true),
       removeSync: sinon.spy(),
     };
-    wrapClean = proxyquire('./wrapClean', {
+    wrapClean = proxyquire('./wrap-clean', {
       'fs-extra': fs,
     });
   });
